APIGateway/requirements.txt
fastapi
grpcio
grpcio-tools
uvicorn
pytest
httpx